test(codegen): extract inline test blocks to sibling tests.rs files
Workspace policy (Makefile:412 lint-tests-layout) forbids inline
#[cfg(test)] mod tests {...} blocks. Extract three of them:

- src/lib.rs::tests → src/tests.rs
- src/client/mod.rs::tests → src/client/tests.rs
- src/client/typescript/render.rs::tests → src/client/typescript/render/tests.rs

No test behavior change; cargo nextest -p fraiseql-codegen → 11/11 pass.
