+++ crates/fraiseql-codegen/src/lib.rs
@@ -40,17 +40,4 @@ pub mod client;
 pub type Generated = BTreeMap<PathBuf, String>;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn generated_is_an_ordered_path_to_content_map() {
-        let mut generated = Generated::new();
-        generated.insert(PathBuf::from("b.ts"), "second".to_string());
-        generated.insert(PathBuf::from("a.ts"), "first".to_string());
-
-        // BTreeMap iteration order is sorted by key, giving deterministic output.
-        let paths: Vec<_> = generated.keys().map(|p| p.to_string_lossy().into_owned()).collect();
-        assert_eq!(paths, vec!["a.ts", "b.ts"]);
-    }
-}
+mod tests;
